chore(classes/resources/specialties) folder renamed
